Break down `contract.py` file
